forge-server/requirements.txt
fastapi
uvicorn[standard]
gitpython
pydantic
aiofiles
